Migrate project form to form-for
