+++ app/controllers/authenticated/projects/new.js
@@ -1,14 +1,17 @@
 import Controller from '@ember/controller';
+import { computed } from '@ember/object';
 import { alias } from '@ember/object/computed';
 import { inject as service } from '@ember/service';
-import ProjectValidations from 'nanowrimo/validations/project';
+import { isEmpty } from '@ember/utils';
 
 export default Controller.extend({
   router: service(),
 
-  ProjectValidations,
-
   project: alias('model'),
+
+  modelErrors: computed('project.genres.[]', function() {
+    return isEmpty(this.get('project.genres')) ? { genres: 'Must select at least one genre' } : null;
+  }),
 
   actions: {
     afterSubmit() {
