+++ app/styles/theme/forms.scss
@@ -1,6 +1,9 @@
 .is-invalid .ember-power-select-trigger {
   border-color: $form-feedback-invalid-color;
-  box-shadow: 0 0 0 0.2rem rgba($form-feedback-invalid-color, 0.25);
+
+  &--active {
+    box-shadow: 0 0 0 0.2rem rgba($form-feedback-invalid-color, 0.25);
+  }
 }
 
 .is-invalid ~ .invalid-feedback {
