+++ app/templates/components/form-for.hbs
@@ -3,6 +3,7 @@
     currentStepIndex=currentStepIndex
     isLastStep=isLastStep
     input=(component 'form-for--input' changeset=changeset hasAttemptedSubmit=hasAttemptedSubmit)
-    select=(component 'form-for--select' changeset=changeset hasAttemptedSubmit=hasAttemptedSubmit)
+    select=( component 'form-for--select' changeset=changeset hasAttemptedSubmit=hasAttemptedSubmit)
+    multiselect=( component 'form-for--multiselect' model=model modelErrors=modelErrors hasAttemptedSubmit=hasAttemptedSubmit)
   )}}
 </form>
