--- app/templates/components/project-form.hbs
@@ -1,21 +1,0 @@
-<form {{action "validateAndSubmit" on="submit"}}>
-  <div class="form-group">
-    {{validated-input
-      errors=changeset.errors
-      hasAttemptedSubmit=hasAttemptedSubmit
-      label="Name"
-      name="name"
-      value=changeset.name
-    }}
-  </div>
-  <div class="form-group">
-    {{validated-genre-select
-      data-test-select-genre=true
-      value=project.genres
-      errorMessage=genreError
-      hasAttemptedSubmit=hasAttemptedSubmit
-    }}
-  </div>
-
-  <input data-test-project-submit type="submit" class="btn btn-primary" value={{buttonLabel}} />
-</form>
