--- app/templates/components/validated-genre-select.hbs
@@ -1,17 +1,0 @@
-<label for={{name}}>Genre(s)</label>
-{{#power-select-multiple-with-create
-  class=selectClasses
-  options=unselectedGenres
-  selected=value
-  searchField="name"
-  showCreatePosition="bottom"
-  showCreateWhen=(action "showCreateWhen")
-  onchange=(action (mut value))
-  oncreate=(action "createGenre")
-  as |genre|
-}}
-  {{genre.name}}
-{{/power-select-multiple-with-create}}
-{{#if showErrorMessage}}
-  <div data-test-error-message class="invalid-feedback">{{errorMessage}}</div>
-{{/if}}
