--- app/templates/components/validated-input.hbs
@@ -1,15 +1,0 @@
-<label for={{name}}>{{label}}</label>
-{{input
-  data-test-validated-input=name
-  autocomplete=autocomplete
-  autofocus=autofocus
-  class=inputClasses
-  name=name
-  focusOut=(action "markBlurred")
-  placeholder=placeholder
-  type=type
-  value=value
-}}
-{{#if showErrorMessage}}
-  <div data-test-error-message class="invalid-feedback">{{errorMessage}}</div>
-{{/if}}
